Move backup status into a dedicated module
